# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HifzMaal API",
        "description": "Islamic Family Finance Management Platform",
        "contact": {
            "email": "support@hifzmaal.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local"
        },
        {
            "url": "https://api.hifzmaal.com",
            "description": "Production"
        }
    ],
    "paths": {
...
        "/api/families/{family}/dashboard": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get dashboard overview",
                "description": "Get comprehensive dashboard data including financial overview, transactions, budgets, bills, savings, and zakat status",
                "operationId": "21c2816fedd9a9eb8b5a17a239551f55",
                "parameters": [
                    {
                        "name": "family",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "financial_overview": {
                                                    "properties": {
                                                        "total_balance": {
                                                            "type": "number"
                                                        },
                                                        "monthly_income": {
                                                            "type": "number"
                                                        },
                                                        "monthly_expense": {
                                                            "type": "number"
                                                        },
                                                        "net_income": {
                                                            "type": "number"
                                                        },
                                                        "currency": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "recent_transactions": {
                                                    "type": "array",
                                                    "items": {}
                                                },
                                                "budget_overview": {
                                                    "type": "object"
                                                },
                                                "bills": {
                                                    "properties": {
                                                        "upcoming": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "object"
                                                            }
                                                        },
                                                        "overdue": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "object"
                                                            }
                                                        },
                                                        "upcoming_count": {
                                                            "type": "integer"
                                                        },
                                                        "overdue_count": {
                                                            "type": "integer"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "savings_overview": {
                                                    "type": "object"
                                                },
                                                "category_expenses": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "monthly_trend": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "zakat_status": {
                                                    "type": "object",
                                                    "nullable": true
                                                },
                                                "pending_approvals": {
                                                    "type": "integer"
                                                },
                                                "family_members": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/families/{family}/insights": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get financial insights and recommendations",
                "description": "Get AI-powered insights, financial health score, and personalized recommendations",
                "operationId": "f8fe7137cb6dea6467f025fc6ab341a0",
                "parameters": [
                    {
                        "name": "family",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "expense_comparison": {
                                                    "properties": {
                                                        "current_month": {
                                                            "type": "number"
                                                        },
                                                        "previous_month": {
                                                            "type": "number"
                                                        },
                                                        "change_percentage": {
                                                            "type": "number"
                                                        },
                                                        "trend": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "top_spending_categories": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "budget_alerts": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "financial_health_score": {
                                                    "type": "integer"
                                                },
                                                "tips": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ],
...
}